+++ packages/loaders/src/index.ts
@@ -1,4 +1,4 @@
-export { loadOmeTiff } from './tiff';
+export { loadOmeTiff, loadMultiTiff } from './tiff';
 export { loadBioformatsZarr, loadOmeZarr } from './zarr';
 
 export { default as TiffPixelSource } from './tiff/pixel-source';